ci(header data): move  headerdata to constant folder
